Assignment1_IBMBLUEMIXFILESTORAGE_4428/requirements.txt
Flask==0.12.3
cf-deployment-tracker==1.0.4

python-swiftclient

pyDES


python-keystoneclient